refactor(app): 拆出全局弹窗类型/优先级助手与内容组件
- 新增 components/globalDialog/globalDialogTypes.ts：队列条目类型、优先级排序助手与开发自检
- 新增 components/globalDialog/DialogContent.tsx：单条弹窗渲染（焦点/键盘/校验/复制逻辑原样搬移）
- GlobalDialog.tsx 579 -> 155 行（收敛为队列管理编排层）
